+++ src/main/scala/com/johnsnowlabs/nlp/annotators/common/TableData.scala
@@ -1,0 +1,3 @@
+package com.johnsnowlabs.nlp.annotators.common
+
+case class TableData(header: Array[String], rows: Array[Array[String]])
